Add retries on docker login

+++ Jenkinsfile
@@ -80,7 +80,18 @@ pipeline {
                 """
 
                 sh """
-                    echo "${DOCKER_CREDS_PSW}" | docker login ${env.DOCKER_REPO} -u ${DOCKER_CREDS_USR} --password-stdin
+                    retries=0
+                    until [ "$retries" -ge 5 ]
+                    do
+                       echo "${DOCKER_CREDS_PSW}" | docker login ${env.DOCKER_REPO} -u ${DOCKER_CREDS_USR} --password-stdin && break
+                       retries=$((retries+1))
+                       echo "docker login failed, sleeping to retry"
+                       sleep 30
+                    done
+                    if [ "$retries" -ge 5]; then
+                       echo "docker login failed after retries"
+                       exit 1
+                    fi
                     rm -rf ${GO_REPO_PATH}/verrazzano
                     mkdir -p ${GO_REPO_PATH}/verrazzano
                     tar cf - . | (cd ${GO_REPO_PATH}/verrazzano/ ; tar xf -)
